complete port connections

+++ src/bridge.js
@@ -1,5 +1,6 @@
 /** allow to use Elm from nodejs...
  */
+const debug = false
 const vm = require('vm')
 const fs = require('fs')
 const proc = require('process')
@@ -8,6 +9,7 @@ const exec = require('child_process').execSync
 function log( o ) {
 	console.log( o )
 }
+if ( debug ) log('############## debug mode is ON ##############')
 
 /** loads Elm compiled javascript
  * and returns Elm object
@@ -21,11 +23,7 @@ function loadElm( path ) {
 
 /** main */
 const Elm = loadElm('./index.js')
-const app = Elm.Main.worker({srccode: 'test'})
-// Elm.Main, {
-// 	srccode: '',
-// 	result: { stdout: '', filename: ''}
-// })
+const app = Elm.Main.worker()
 
 console.log('Starting elm-doctest ...')
 if ( proc.argv.length != 3 ) {
@@ -35,24 +33,30 @@ if ( proc.argv.length != 3 ) {
 const elmfile = proc.argv[ 2 ]
 try {
 	const elmsrc = fs.readFileSync( elmfile, 'utf8')
-	app.ports.srccode.send( elmsrc )
+	setTimeout(() => {  // for some reason, port does not work without delay.
+		app.ports.srccode.send( elmsrc )
+	}, 1)
 } catch(e) {
 	log(e)
 	process.exit( 1 )
 }
 
 app.ports.evaluate.subscribe(function( resource ) {
-	log('----------- evaluate called.')
+	if ( debug ) {
+		log('----------- evaluate called.')
+		log(resource)
+	}
 	if ( resource.src.length == 0 ) return
 	// log('writing temporary source into file...')
 	fs.writeFileSync('./src/DoctestTempModule__.elm', resource.src )
 	const stdout = exec('elm-repl', { input: resource.runner, encoding: 'utf8'})
-	// log( stdout )
+	if ( debug ) log( stdout )
 	const match = stdout.match(/^> (.+)/gm)
 	if ( !match ) return []
 	const resultStr = match[0].replace(/[^"]*(".+")( : .*)?/, '$1')
+	if ( debug ) log( resultStr )
 	app.ports.result.send({ stdout: JSON.parse( resultStr ), filename: elmfile })
-	if ( fs.existsSync('./src/DoctestTempModule__.elm'))
+	if ( !debug && fs.existsSync('./src/DoctestTempModule__.elm'))
 		fs.unlinkSync('./src/DoctestTempModule__.elm')
 })
